third_party/WebKit/LayoutTests/fast/speechsynthesis/speech-synthesis-speak-invalid-argument-throws.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="console"></div>

<script>

    if (window.internals)
        window.internals.enableMockSpeechSynthesizer();

    description("This tests that passing an argument other than a SpeechSynthesisUtterance to speechSynthesis.speak throws a TypeError exception.");

    shouldThrow("speechSynthesis.speak()", '"TypeError: Failed to execute \'speak\' on \'SpeechSynthesis\': 1 argument required, but only 0 present."');
    shouldThrow("speechSynthesis.speak(0)", '"TypeError: Invalid utterance argument"');
    shouldThrow("speechSynthesis.speak('')", '"TypeError: Invalid utterance argument"');
    shouldThrow("speechSynthesis.speak(document.body)", '"TypeError: Invalid utterance argument"');
    shouldThrow("speechSynthesis.speak({})", '"TypeError: Invalid utterance argument"');
    shouldThrow("speechSynthesis.speak((new SpeechSynthesisUtterance()).text = 'hello')", '"TypeError: Invalid utterance argument"');

</script>

<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
