third_party/WebKit/LayoutTests/fast/speechsynthesis/speech-synthesis-speak-invalid-argument-throws-expected.txt

This tests that passing an argument other than a SpeechSynthesisUtterance to speechSynthesis.speak throws a TypeError exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS speechSynthesis.speak() threw exception TypeError: Failed to execute 'speak' on 'SpeechSynthesis': 1 argument required, but only 0 present..
PASS speechSynthesis.speak(0) threw exception TypeError: Failed to execute 'speak' on 'SpeechSynthesis': Invalid utterance argument.
PASS speechSynthesis.speak('') threw exception TypeError: Failed to execute 'speak' on 'SpeechSynthesis': Invalid utterance argument.
PASS speechSynthesis.speak(document.body) threw exception TypeError: Failed to execute 'speak' on 'SpeechSynthesis': Invalid utterance argument.
PASS speechSynthesis.speak({}) threw exception TypeError: Failed to execute 'speak' on 'SpeechSynthesis': Invalid utterance argument.
PASS speechSynthesis.speak((new SpeechSynthesisUtterance()).text = 'hello') threw exception TypeError: Failed to execute 'speak' on 'SpeechSynthesis': Invalid utterance argument.
PASS successfullyParsed is true

TEST COMPLETE
